Extract ManageAccountButton and use in new SignIn component
